Create NBulbPuzzle.java
There are N bulbs, numbered 1 to N, arranged in a row. The first bulb is plugged into the power socket and each successive bulb is connected to the previous one(1-> 2->3...).
Initially, all the bulbs are turned off. At moment K(for K from 0 to N-1), we turn on A[K]-th bulb. A bulb shines if it is on and all the previous bulbs are turned on too.

Write a function solution that given an array A of N different integers from 1 to N, returns the number of moments for which every turned on bulb shines.